better demo for gated content
